added design component
