+++ package.json
@@ -27,7 +27,6 @@
     "react-native-safe-area-context": "3.2.0",
     "react-native-screens": "~3.4.0",
     "react-native-web": "~0.13.12",
-    "react-native-webview": "11.6.2",
     "react-navigation-collapsible": "^5.10.2"
   },
   "devDependencies": {
